cmd/atlascmd/migrate: introduce SchemaChanger interface (#831)
* cmd/atlascmd/migrate: introduce SchemaChanger interface

Ent based schema revisions are stored in its own database schema (except SQLite). The responsibility for parsing and editing connection URLs lay at the drivers. Ent based revisions introduce a new interface for the drivers to implement in order to alter the originally given connection URL to connect to another schema to store revisions table in.

* fix tests

* CR

* rebase and test fix

* postgres has no error on non-existing database -> make sure schema exists even on successful connection attempt

* CR