internal: database connections for a test are established in the test… (#1031)
* internal: database connections for a test are established in the tests itself.

With this changes it os no longer required to have all integration containers running if one only works on one dialect or version. This works by only then spinning up a connection to a database if it is requested the first time. Subsequent calls re-use the existing connection.

If a connection for a container is requested and the container is not running, there still will be an error and the test will fail.

CI has been changed to respect those changes.

* internal/integration: add README to explain how to selectively run tests

* apply CR